requirements0605.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.7
django-bootstrap4==0.0.7
django-crispy-forms==1.7.2
django-formtools==2.1
django-generate-secret-key==1.0.2
django-multiselectfield==0.1.8
dohq-youtrack==0.2.1.60
httplib2==0.12.0
idna==2.8
Pillow==5.4.1
pytz==2018.9
requests==2.21.0
sqlite3==0.0.0
urllib3==1.24.1
XlsxWriter==1.1.4
